network service added

+++ swiftyCompanion/Model.swift
@@ -37,22 +37,26 @@ struct User: Decodable {
     var poolYear: String
     var wallet: Int
     var login: String
+//    var projects42: [Projects]
+//    var skills42: [SkillsInfo]
     var projectsAll: [Projects]
     var cursusAll: [CursusInfo]
-    
+
     struct Projects: Decodable {
-//        var id: Int
         var finalMark: Int?
         var status: String
         var cursusId: [Int]
         var project: ProjectInfo
+//        var name: String
+//        var slug: String
         
         enum CodingKeys : String, CodingKey {
-//            case id
             case finalMark = "final_mark"
             case status
             case cursusId = "cursus_ids"
             case project = "project"
+//            case name
+//            case slug
         }
     }
     
@@ -63,7 +67,7 @@ struct User: Decodable {
         
     struct CursusInfo: Decodable {
         var beginAt: String
-        var level: Float
+        var level: Double
         var cursus: Cursus
         var skills: [SkillsInfo]
         
@@ -85,55 +89,21 @@ struct User: Decodable {
         }
     
     enum CodingKeys : String, CodingKey {
-      case id
-      case email
-      case firstName = "first_name"
-      case lastName = "last_name"
-      case displayName = "displayname"
-      case imageURL = "image_url"
-      case correctionPoint = "correction_point"
-      case poolMonth = "pool_month"
-      case poolYear = "pool_year"
-      case wallet
-      case login
-      case projectsAll = "projects_users"
-      case cursusAll = "cursus_users"
+        case id
+        case email
+        case firstName = "first_name"
+        case lastName = "last_name"
+        case displayName = "displayname"
+        case imageURL = "image_url"
+        case correctionPoint = "correction_point"
+        case poolMonth = "pool_month"
+        case poolYear = "pool_year"
+        case wallet
+        case login
+
+//        case projects42
+//        case skills42
+        case projectsAll = "projects_users"
+        case cursusAll = "cursus_users"
     }
 }
-
-//
-//struct Projects: Decodable {
-//    var finalMark: Int?
-//    var status: String
-//    var project: ProjectInfo?
-//
-//    enum CodingKeys : String, CodingKey {
-//      case finalMark = "final_mark"
-//      case status
-//      case project = "project"
-//
-//    }
-//}
-//
-//struct ProjectInfo: Decodable {
-//    var name: String
-//    var slug: String
-//}
-//
-//struct CursusInfo: Decodable {
-//    var beginAt: String
-//    var level: Float
-//    var skills: [SkillsInfo?]
-//
-//    enum CodingKeys : String, CodingKey {
-//      case beginAt = "begin_at"
-//      case level
-//      case skills = "skills"
-//
-//    }
-//}
-//
-//struct SkillsInfo: Decodable {
-//    var name: String
-//    var level: Double
-//}
